+++ src/commands/impls/keys.rs
@@ -148,6 +148,14 @@ pub async fn expire_at<C: ClientLike>(client: &C, key: RedisKey, timestamp: i64)
   protocol_utils::frame_to_results(frame)
 }
 
+pub async fn expire_time<C: ClientLike>(client: &C, key: RedisKey) -> Result<RedisValue, RedisError> {
+  one_arg_value_cmd(client, RedisCommandKind::ExpireTime, key.into()).await
+}
+
+pub async fn pexpire_time<C: ClientLike>(client: &C, key: RedisKey) -> Result<RedisValue, RedisError> {
+  one_arg_value_cmd(client, RedisCommandKind::PexpireTime, key.into()).await
+}
+
 pub async fn pexpire<C: ClientLike>(
   client: &C,
   key: RedisKey,
@@ -411,6 +419,10 @@ pub async fn watch<C: ClientLike>(client: &C, keys: MultipleKeys) -> Result<(), 
 }
 
 ok_cmd!(unwatch, Unwatch);
+
+pub async fn r#type<C: ClientLike>(client: &C, key: RedisKey) -> Result<RedisValue, RedisError> {
+  one_arg_value_cmd(client, RedisCommandKind::Type, key.into()).await
+}
 
 pub async fn lcs<C: ClientLike>(
   client: &C,